feat(02): calculate number of digits in a uint64
Because I think I can calculate the number repeated patterns with a
given number of digits, I need to be able to calculate the number of
digits in a number to divide up the work.

The choice of uint64 is because the input values exceed 4 billion, and I
don't need negative numbers.