feat(05): create span merge function
I am thinking about doing a binary search in the ranges, which will
require sorting the ranges. After sorting, I can also easily merge all
the overlapping range, so do that too so that there are fewer ranges to
check.

I called ranges span instead of ranges because range is a keyword in go,
so I cannot name a variable range.